Agents/requirements.txt
faiss-cpu
pandas
numpy
sentence-transformers
scikit-learn
neo4j
tqdm
python-dotenv
